requirements-render.txt
# Minimal requirements for Render deployment
# This is a fallback requirements file with minimal versions

# Core framework
streamlit==1.28.0

# Document processing
PyPDF2==3.0.1
python-docx==0.8.11

# Basic AI/ML 
numpy==1.21.6
scikit-learn==1.2.2
openai==1.3.9

# Lightweight ML dependencies
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Database
sqlalchemy==1.4.46
psycopg2-binary==2.9.5

# Basic utilities
pandas==1.5.3
pillow==9.4.0
requests==2.28.2